:hammer: Move utils/data.py to datasets/utils.py
