+++ torch_uncertainty/datamodules/segmentation/cityscapes.py
@@ -9,8 +9,8 @@ from torchvision.transforms import v2
 
 from torch_uncertainty.datamodules import TUDataModule
 from torch_uncertainty.datasets.segmentation import Cityscapes
+from torch_uncertainty.datasets.utils import create_train_val_split
 from torch_uncertainty.transforms import RandomRescale
-from torch_uncertainty.utils import create_train_val_split
 
 
 class CityscapesDataModule(TUDataModule):
