+++ torch_uncertainty/utils/__init__.py
@@ -1,7 +1,6 @@
 # ruff: noqa: F401
 from .checkpoints import get_version
 from .cli import TULightningCLI
-from .data import TTADataset, create_train_val_split
 from .hub import load_hf
 from .misc import csv_writer, plot_hist
 from .trainer import TUTrainer
